Change project structure to match the new Architecture
The project will be migrated from MVP to MVVM + Clean Arch.
Therefore, it was necessary to change the project structure
to match the new architecture.
