+++ app/src/main/AndroidManifest.xml
@@ -5,15 +5,14 @@
     <uses-permission android:name="android.permission.ACCESS_NOTIFICATION_POLICY"/>
 
     <application
-        android:allowBackup="true"
         android:icon="@mipmap/ic_app"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_app"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.navigationAPI.NavigationAPIActivity" />
+        <activity android:name=".framework.navigationAPI.NavigationAPIActivity" />
         <activity
-            android:name=".ui.splash.SplashActivity"
+            android:name=".presentation.feature.splash.SplashActivity"
             android:noHistory="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
