+++ app/src/main/java/com/tab/mypomodoroapp/framework/navigationAPI/NavigationAPIContract.kt
@@ -1,4 +1,4 @@
-package com.tab.mypomodoroapp.ui.navigationAPI
+package com.tab.mypomodoroapp.framework.navigationAPI
 
 import androidx.fragment.app.Fragment
