+++ app/src/main/java/com/tab/mypomodoroapp/presentation/feature/fragment/startPomodoro/StartPomodoroContract.kt
@@ -1,6 +1,6 @@
-package com.tab.mypomodoroapp.ui.fragment.startPomodoro
+package com.tab.mypomodoroapp.presentation.feature.fragment.startPomodoro
 
-import com.tab.mypomodoroapp.ui.navigationAPI.NavigationAPIContract
+import com.tab.mypomodoroapp.framework.navigationAPI.NavigationAPIContract
 
 interface StartPomodoroContract {
     interface StartPomodoroPresenter {
