+++ app/src/main/res/layout/activity_initial.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorAppBackground"
-    tools:context=".ui.navigationAPI.NavigationAPIActivity">
+    tools:context=".framework.navigationAPI.NavigationAPIActivity">
 
     <LinearLayout
         android:id="@+id/fragment_container"
