+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorAppBackground"
-    tools:context=".ui.splash.SplashActivity">
+    tools:context=".presentation.feature.splash.SplashActivity">
 
     <ImageView
         android:id="@+id/imageView"
