+++ app/src/main/res/layout/fragment_count_down.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.countdown.CountDownFragment">
+    tools:context=".presentation.feature.fragment.countdown.CountDownFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
