+++ app/src/main/res/layout/fragment_pomodoro_follow_up.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.pomodoro.PomodoroFragment">
+    tools:context=".presentation.feature.fragment.pomodoro.PomodoroFragment">
