+++ app/src/main/res/layout/fragment_start_pomodoro.xml
@@ -6,7 +6,7 @@ android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:background="@color/colorAppBackground"
 android:padding="16dp"
-tools:context=".ui.fragment.startPomodoro.StartPomodoroFragment">
+tools:context=".presentation.feature.fragment.startPomodoro.StartPomodoroFragment">
 
 <ImageView
     android:id="@+id/imageView2"
