+++ core/src/main/java/com/tab/core/model/Param.kt
@@ -1,0 +1,3 @@
+package com.tab.core.model
+
+data class Param<Input>(val input: Input)
